+++ CHANGELOG.md
@@ -2,3 +2,13 @@
 
 - Changes console prompts to Spectre.Console
 - Fixes CHANGELOG concatenation
+
+## 1.2.0
+
+- Adds All Projects selection
+- Adds check for csproj missing Version element
+
+## 1.3.0
+
+- Adds explain command
+- Adds validate command